+++ src/com/landawn/abacus/util/BooleanList.java
@@ -25,6 +25,7 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 
+import com.landawn.abacus.util.Try.Function;
 import com.landawn.abacus.util.function.BiConsumer;
 import com.landawn.abacus.util.function.BiFunction;
 import com.landawn.abacus.util.function.IntFunction;
@@ -1390,6 +1391,11 @@ public final class BooleanList extends PrimitiveList<Boolean, boolean[], Boolean
         checkFromToIndex(fromIndex, toIndex);
 
         return Stream.of(elementData, fromIndex, toIndex);
+    }
+
+    @Override
+    public <R, E extends Exception> Optional<R> ifNotEmpty(Function<? super BooleanList, R, E> func) throws E {
+        return isEmpty() ? Optional.<R> empty() : Optional.of(func.apply(this));
     }
 
     @Override