+++ src/com/landawn/abacus/util/CharList.java
@@ -25,6 +25,7 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 
+import com.landawn.abacus.util.Try.Function;
 import com.landawn.abacus.util.function.BiConsumer;
 import com.landawn.abacus.util.function.BiFunction;
 import com.landawn.abacus.util.function.IntFunction;
@@ -1518,6 +1519,11 @@ public final class CharList extends PrimitiveList<Character, char[], CharList> {
         checkFromToIndex(fromIndex, toIndex);
 
         return CharStream.of(elementData, fromIndex, toIndex);
+    }
+
+    @Override
+    public <R, E extends Exception> Optional<R> ifNotEmpty(Function<? super CharList, R, E> func) throws E {
+        return isEmpty() ? Optional.<R> empty() : Optional.of(func.apply(this));
     }
 
     @Override