+++ src/com/landawn/abacus/util/FloatList.java
@@ -26,6 +26,7 @@ import java.util.Map;
 import java.util.Random;
 import java.util.Set;
 
+import com.landawn.abacus.util.Try.Function;
 import com.landawn.abacus.util.function.BiConsumer;
 import com.landawn.abacus.util.function.BiFunction;
 import com.landawn.abacus.util.function.IntFunction;
@@ -1491,6 +1492,11 @@ public final class FloatList extends PrimitiveList<Float, float[], FloatList> {
         checkFromToIndex(fromIndex, toIndex);
 
         return FloatStream.of(elementData, fromIndex, toIndex);
+    }
+
+    @Override
+    public <R, E extends Exception> Optional<R> ifNotEmpty(Function<? super FloatList, R, E> func) throws E {
+        return isEmpty() ? Optional.<R> empty() : Optional.of(func.apply(this));
     }
 
     @Override