+++ src/com/landawn/abacus/util/PrimitiveList.java
@@ -278,6 +278,8 @@ public abstract class PrimitiveList<B, A, L extends PrimitiveList<B, A, L>> impl
 
     public abstract Multiset<B> toMultiset(final int fromIndex, final int toIndex, final IntFunction<Multiset<B>> supplier);
 
+    public abstract <R, E extends Exception> Optional<R> ifNotEmpty(Try.Function<? super L, R, E> func) throws E;
+
     public abstract <R, E extends Exception> R apply(Try.Function<? super L, R, E> func) throws E;
 
     public abstract <E extends Exception> void accept(Try.Consumer<? super L, E> action) throws E;